server/requirements.txt
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
django-mptt==0.12.0
drf-spectacular==0.21.1 
psycopg2-binary>=2.9.1,<2.9.2
djangorestframework-simplejwt==4.6.0
django-jazzmin==2.4.8
djoser==2.0.5
django-redis>=5.2.0
boto3>=1.15.0
django-storages==1.12.3
Pillow==9.0.0
black==22.12.0
flake8==6.0.0
-e ./drf-instamojo
